Tweak OhcCache
